Create integration test for sortable entity.
